services/similarity/requirements.txt
mysql-connector-python
sentence-transformers
torch
transformers==4.35.2
huggingface_hub==0.16.4